backend/requirements.txt
fastapi
uvicorn
duckdb
anthropic
python-dotenv
ddgs
